craw_agent/pyproject.toml
[project]
name = "craw-agent"
version = "0.1.0"
description = "Weekly Report Generation Crew using CrewAI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.11.0",
    "crewai-tools>=0.78.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
weekly-report-crew = "weekly_report_crew.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
